+++ modules/core/src/main/scala/moe/brianhsu/live2d/adapter/gateway/avatar/CubismModelBackend.scala
@@ -8,9 +8,9 @@ import moe.brianhsu.live2d.enitiy.core.memory.MemoryInfo
 import moe.brianhsu.live2d.enitiy.core.types.{CPointerToMoc, CPointerToModel, ModelAlignment}
 import moe.brianhsu.live2d.enitiy.model
 import moe.brianhsu.live2d.enitiy.model.drawable.Drawable.ColorFetcher
-import moe.brianhsu.live2d.enitiy.model.drawable.{ConstantFlags, Drawable, DrawableColor, DynamicFlags, VertexInfo}
+import moe.brianhsu.live2d.enitiy.model.drawable._
 import moe.brianhsu.live2d.enitiy.model.{MocInfo, ModelCanvasInfo, Parameter, Part}
-import moe.brianhsu.live2d.exception.{DrawableInitException, MocNotRevivedException, ParameterInitException, PartInitException, TextureSizeMismatchException}
+import moe.brianhsu.live2d.exception._
 
 import scala.util.Try