+++ modules/core/src/test/scala/moe/brianhsu/live2d/enitiy/avatar/effect/impl/FaceTrackingFeature.scala
@@ -62,7 +62,7 @@ class FaceTrackingFeature extends AnyFeatureSpec with GivenWhenThen with Matcher
 
   private val DefaultFaceTrackingTaps = TrackingTaps(3, 4, 5, 6, 7, 8, 9, 10, 11, 11)
 
-  case class FixedFaceTracking(currentTrackingNoes: List[TrackingNode]) extends FaceTracking(DefaultFaceTrackingTaps) {
+  class FixedFaceTracking(currentTrackingNoes: List[TrackingNode]) extends FaceTracking(DefaultFaceTrackingTaps) {
     this.trackingNoes = currentTrackingNoes
   }