+++ modules/core/src/test/scala/moe/brianhsu/live2d/enitiy/updater/ModelUpdaterFeature.scala
@@ -69,6 +69,11 @@ class ModelUpdaterFeature extends AnyFeatureSpec with GivenWhenThen with Matcher
       val mockedParameterId1 = stub[Parameter]
       val mockedParameterId2 = stub[Parameter]
       val model = stub[Live2DModel]
+      val mockedParameters = Map(
+        "id1" -> mockedParameterId1,
+        "id2" -> mockedParameterId2,
+      )
+      (() => model.parameters).when().returns(mockedParameters)
       (model.parameterWithFallback _).when("id1").returns(mockedParameterId1)
       (model.parameterWithFallback _).when("id2").returns(mockedParameterId2)
 
@@ -123,5 +128,97 @@ class ModelUpdaterFeature extends AnyFeatureSpec with GivenWhenThen with Matcher
         (mockedPart1.opacity_= _).verify(0.4f).once()
       }
     }
+  }
+
+  Feature("Convert to old parameter id format") {
+    Scenario("Update ParamTere should change to update PARAM_CHEEK") {
+      Given("a stubbed model with some stubbed parameters")
+      val mockedParameterId1 = stub[Parameter]
+      val mockedParameterId2 = stub[Parameter]
+      val model = stub[Live2DModel]
+      val mockedParameters = Map(
+        "PARAM_ANGLE_X" -> mockedParameterId1,
+        "PARAM_CHEEK" -> mockedParameterId2,
+      )
+      (() => model.parameters).when().returns(mockedParameters)
+
+      And("a ModelUpdater based on that model")
+      val updater = new ModelUpdater(model)
+
+      When("executeOperations with a operation list to that model")
+      updater.executeOperations(
+        List(
+          ParameterValueUpdate("ParamTere", 0.1f, 0.2f),
+        )
+      )
+
+      Then("it should delegated it to parameter correctly")
+      inSequence {
+        (mockedParameterId2.update _).verify(0.1f, 0.2f).once()
+      }
+
+    }
+
+    Scenario("Update ParamAngleX should change to update PARAM_ANGLE_X") {
+      Given("a stubbed model with some stubbed parameters")
+      val mockedParameterId1 = stub[Parameter]
+      val mockedParameterId2 = stub[Parameter]
+      val model = stub[Live2DModel]
+      val mockedParameters = Map(
+        "PARAM_ANGLE_X" -> mockedParameterId1,
+        "PARAM_CHEEK" -> mockedParameterId2,
+      )
+      (() => model.parameters).when().returns(mockedParameters)
+
+      And("a ModelUpdater based on that model")
+      val updater = new ModelUpdater(model)
+
+      When("executeOperations with a operation list to that model")
+      updater.executeOperations(
+        List(
+          ParameterValueUpdate("ParamAngleX", 0.1f, 0.2f),
+        )
+      )
+
+      Then("it should delegated it to parameter correctly")
+      inSequence {
+        (mockedParameterId1.update _).verify(0.1f, 0.2f).once()
+      }
+    }
+
+    Scenario("The operation contains multiple operations that access parts") {
+      Given("a stubbed model with some stubbed parts")
+      val mockedPart1 = stub[Part]
+      val mockedPart2 = stub[Part]
+      val parts = Map("id1" -> mockedPart1, "id2" -> mockedPart2)
+      val mockedParameters = Map(
+        "PARAM_ANGLE_X" -> stub[Parameter],
+      )
+      val model = stub[Live2DModel]
+      (() => model.parts).when().returns(parts)
+      (() => model.parameters).when().returns(mockedParameters)
+
+      And("a ModelUpdater based on that model")
+      val updater = new ModelUpdater(model)
+
+      When("executeOperations with a operation list to that model")
+      updater.executeOperations(
+        List(
+          PartOpacityUpdate("id1", 0.1f),
+          PartOpacityUpdate("id2", 0.2f),
+          PartOpacityUpdate("id2", 0.3f),
+          PartOpacityUpdate("id1", 0.4f),
+        )
+      )
+
+      Then("it should delegated it to parameter correctly without change it's name")
+      inSequence {
+        (mockedPart1.opacity_= _).verify(0.1f).once()
+        (mockedPart2.opacity_= _).verify(0.2f).once()
+        (mockedPart2.opacity_= _).verify(0.3f).once()
+        (mockedPart1.opacity_= _).verify(0.4f).once()
+      }
+    }
+
   }
 }