+++ CHANGELOG.md
@@ -7,6 +7,9 @@ CHANGELOG
 ### Core API
 
 - Added Features
-  - OpenSeeFace UDP data reader API.
+  - `UDPOpenSeeFaceDataReader' API.
+  - `FaceTracking` / `OpenSeeFaceTracking` effect.
+  - `Live2DModel` has new `isOldParameterId` attribute indicates if it use old parameter id form.
+  - `ModelUpdater` now will convert new parameter id format to old format when using old model.
 - Breaking API Changes
   - Remove the adapter dependency in AvatarMotion factory method.